@@ -10,9 +10,9 @@
     "antd": "^5.24.2",
     "axios": "^1.8.1",
     "chart.js": "^4.4.8",
-    "react": "^17.0.1",
+    "react": "^18.3.1",
     "react-chartjs-2": "^5.3.0",
-    "react-dom": "^17.0.1",
+    "react-dom": "^18.3.1",
     "react-icons": "^5.5.0",
     "react-router-dom": "^6.29.0",
     "react-scripts": "^5.0.1",
