@@ -3,60 +3,70 @@ export const aptitudeQuestions = [
         "question": "A train 150 meters long is running at a speed of 54 km/hr. How long will it take to cross a pole?",
         "options": ["5 sec", "8 sec", "10 sec", "12 sec"],
         "correctOption": 2,
-        "points": 10
+        "points": 10,
+        "category": "aptitude"
     },
     {
         "question": "The sum of two numbers is 72, and their difference is 24. What is the larger number?",
         "options": ["24", "36", "48", "60"],
         "correctOption": 2,
-        "points": 10
+        "points": 10,
+        "category": "aptitude"
     },
     {
         "question": "A shopkeeper gives a 20% discount on an article that originally costs ₹500. What is the final price?",
         "options": ["₹400", "₹450", "₹480", "₹520"],
         "correctOption": 0,
-        "points": 10
+        "points": 10,
+        "category": "aptitude"
     },
     {
         "question": "A and B can complete a task in 20 days. A alone can do it in 30 days. In how many days can B complete it alone?",
         "options": ["40", "50", "60", "70"],
         "correctOption": 2,
-        "points": 10
+        "points": 10,
+        "category": "aptitude"
     },
     {
         "question": "A train moving at 72 km/hr crosses a 180-meter platform in 18 seconds. What is the length of the train?",
         "options": ["120m", "150m", "180m", "200m"],
         "correctOption": 2,
-        "points": 10
+        "points": 10,
+        "category": "aptitude"
     },
     {
         "question": "A man spends 40% of his salary on rent, 30% on food, and saves the remaining ₹12,000. What is his total salary?",
         "options": ["₹40,000", "₹50,000", "₹60,000", "₹70,000"],
         "correctOption": 1,
-        "points": 10
+        "points": 10,
+        "category": "aptitude"
     },
     {
         "question": "The ratio of ages of A and B is 3:5. After 6 years, their ratio becomes 5:7. What is A’s current age?",
         "options": ["12", "15", "18", "21"],
         "correctOption": 2,
-        "points": 10
+        "points": 10,
+        "category": "aptitude"
     },
     {
         "question": "A sum of ₹5,000 is deposited at 10% per annum simple interest. What is the total amount after 4 years?",
         "options": ["₹6,500", "₹7,000", "₹7,500", "₹8,000"],
         "correctOption": 1,
-        "points": 10
+        "points": 10,
+        "category": "aptitude"
     },
     {
         "question": "Two pipes A and B can fill a tank in 20 and 30 minutes respectively. If both are opened together, how long will it take to fill the tank?",
         "options": ["10 min", "12 min", "15 min", "18 min"],
         "correctOption": 1,
-        "points": 10
+        "points": 10,
+        "category": "aptitude"
     },
     {
         "question": "A person travels 120 km at a speed of 40 km/hr and then another 120 km at 60 km/hr. What is his average speed?",
         "options": ["45 km/hr", "48 km/hr", "50 km/hr", "55 km/hr"],
         "correctOption": 1,
-        "points": 10
+        "points": 10,
+        "category": "aptitude"
     }
 ];
