@@ -3,13 +3,15 @@ export const communicationQuestions = [
         "question": "Which word means the same as 'Eloquent'?",
         "options": ["Silent", "Persuasive", "Confused", "Weak"],
         "correctOption": 1,
-        "points": 10
+        "points": 10,
+        "category": "communication"
     },
     {
         "question": "Choose the correct synonym for 'Resilient'.",
         "options": ["Fragile", "Flexible", "Rigid", "Lazy"],
         "correctOption": 1,
-        "points": 10
+        "points": 10,
+        "category": "communication"
     },
     {
         "question": "Which sentence is grammatically correct?",
@@ -20,19 +22,22 @@ export const communicationQuestions = [
             "She not like coffee."
         ],
         "correctOption": 2,
-        "points": 10
+        "points": 10,
+        "category": "communication"
     },
     {
         "question": "What is the antonym of 'Ambiguous'?",
         "options": ["Unclear", "Definite", "Vague", "Doubtful"],
         "correctOption": 1,
-        "points": 10
+        "points": 10,
+        "category": "communication"
     },
     {
         "question": "Identify the word that means 'brief and to the point'.",
         "options": ["Verbose", "Concise", "Elaborate", "Lengthy"],
         "correctOption": 1,
-        "points": 10
+        "points": 10,
+        "category": "communication"
     },
     {
         "question": "Which sentence has the correct subject-verb agreement?",
@@ -43,7 +48,8 @@ export const communicationQuestions = [
             "The group of students have been playing outside."
         ],
         "correctOption": 1,
-        "points": 10
+        "points": 10,
+        "category": "communication"
     },
     {
         "question": "Find the correct idiom meaning 'to take on more responsibility than one can handle'.",
@@ -54,13 +60,15 @@ export const communicationQuestions = [
             "Jump on the bandwagon"
         ],
         "correctOption": 1,
-        "points": 10
+        "points": 10,
+        "category": "communication"
     },
     {
         "question": "Which word best fits in the blank? 'The CEO delivered a ___ speech at the conference.'",
         "options": ["Verbose", "Persuasive", "Tedious", "Confusing"],
         "correctOption": 1,
-        "points": 10
+        "points": 10,
+        "category": "communication"
     },
     {
         "question": "What is the meaning of 'turning a blind eye'?",
@@ -71,12 +79,14 @@ export const communicationQuestions = [
             "Being unaware of something"
         ],
         "correctOption": 1,
-        "points": 10
+        "points": 10,
+        "category": "communication"
     },
     {
         "question": "Which word means the same as 'Plausible'?",
         "options": ["Unbelievable", "Reasonable", "Fake", "False"],
         "correctOption": 1,
-        "points": 10
+        "points": 10,
+        "category": "communication"
     }
 ];
