@@ -3,25 +3,29 @@ export const frontendQuestions = [
         "question": "Which HTML tag is used to define an unordered list?",
         "options": ["<ol>", "<ul>", "<li>", "<list>"],
         "correctOption": 1,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     },
     {
         "question": "Which CSS property is used to change the text color of an element?",
         "options": ["background-color", "text-style", "color", "font-color"],
         "correctOption": 2,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     },
     {
         "question": "Which JavaScript function is used to select an element by its ID?",
         "options": ["document.querySelector()", "document.getElementById()", "document.selectId()", "document.getElement()"],
         "correctOption": 1,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     },
     {
         "question": "Which of the following is not a valid JavaScript data type?",
         "options": ["Number", "Boolean", "Float", "Undefined"],
         "correctOption": 2,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     },
     {
         "question": "Which of the following is correct about React?",
@@ -32,13 +36,15 @@ export const frontendQuestions = [
             "React doesn't use components"
         ],
         "correctOption": 1,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     },
     {
         "question": "Which lifecycle method in React is used for API calls?",
         "options": ["componentDidMount", "componentWillUnmount", "useEffect", "render"],
         "correctOption": 2,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     },
     {
         "question": "What does the 'position: absolute' property do in CSS?",
@@ -49,13 +55,15 @@ export const frontendQuestions = [
             "Positions the element relative to its default position"
         ],
         "correctOption": 0,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     },
     {
         "question": "Which HTTP method is used to update an existing resource?",
         "options": ["GET", "POST", "PUT", "DELETE"],
         "correctOption": 2,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     },
     {
         "question": "What is the purpose of the useState() hook in React?",
@@ -66,7 +74,8 @@ export const frontendQuestions = [
             "To create CSS styles dynamically"
         ],
         "correctOption": 0,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     },
     {
         "question": "Which of the following is a valid way to create a function in JavaScript?",
@@ -77,6 +86,7 @@ export const frontendQuestions = [
             "All of the above"
         ],
         "correctOption": 3,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     }
 ];
