@@ -3,60 +3,70 @@ export const javaQuestions = [
         "question": "What is the default value of an int variable in Java?",
         "options": ["0", "null", "undefined", "None"],
         "correctOption": 0,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     },
     {
         "question": "Which keyword is used to define a constant variable in Java?",
         "options": ["constant", "final", "static", "const"],
         "correctOption": 1,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     },
     {
         "question": "Which Java collection allows duplicate values?",
         "options": ["Set", "List", "Map", "None"],
         "correctOption": 1,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     },
     {
         "question": "Which method is used to start a thread in Java?",
         "options": ["run()", "execute()", "start()", "launch()"],
         "correctOption": 2,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     },
     {
         "question": "Which OOP principle allows method overloading in Java?",
         "options": ["Polymorphism", "Encapsulation", "Abstraction", "Inheritance"],
         "correctOption": 0,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     },
     {
         "question": "Which exception is thrown when an array is accessed out of its bounds?",
         "options": ["IOException", "NullPointerException", "ArrayIndexOutOfBoundsException", "IllegalArgumentException"],
         "correctOption": 2,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     },
     {
         "question": "Which method is called automatically when an object is garbage collected?",
         "options": ["clean()", "finalize()", "delete()", "destroy()"],
         "correctOption": 1,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     },
     {
         "question": "Which Java access modifier allows a variable to be accessible within the same package?",
         "options": ["private", "protected", "public", "default"],
         "correctOption": 3,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     },
     {
         "question": "Which Java class is used to take input from the user?",
         "options": ["System", "Scanner", "BufferedReader", "Reader"],
         "correctOption": 1,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     },
     {
         "question": "What does JVM stand for?",
         "options": ["Java Virtual Machine", "Java Variable Manager", "Java Visual Model", "Java Verified Memory"],
         "correctOption": 0,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     }
 ];
