@@ -3,60 +3,70 @@ export const nodejsQuestions = [
         "question": "Which command is used to initialize a new Node.js project?",
         "options": ["node init", "npm start", "npm init", "create node"],
         "correctOption": 2,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     },
     {
         "question": "Which module is used to create an HTTP server in Node.js?",
         "options": ["http", "express", "server", "net"],
         "correctOption": 0,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     },
     {
         "question": "Which command installs Express.js?",
         "options": ["npm install express", "node install express", "express install", "install express"],
         "correctOption": 0,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     },
     {
         "question": "Which function is used to read a file in Node.js?",
         "options": ["readFile()", "fs.read()", "fs.readFile()", "file.read()"],
         "correctOption": 2,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     },
     {
         "question": "Which object is used to handle file paths in Node.js?",
         "options": ["fs", "path", "os", "url"],
         "correctOption": 1,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     },
     {
         "question": "Which global object in Node.js is used to schedule asynchronous operations?",
         "options": ["setInterval", "setTimeout", "process", "nextTick"],
         "correctOption": 3,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     },
     {
         "question": "Which statement is used to import a module in Node.js (CommonJS)?",
         "options": ["import", "require", "include", "load"],
         "correctOption": 1,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     },
     {
         "question": "Which method is used to create a new instance of an Express app?",
         "options": ["express.init()", "express.create()", "express()", "express.new()"],
         "correctOption": 2,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     },
     {
         "question": "Which Node.js module allows working with the file system?",
         "options": ["path", "fs", "util", "os"],
         "correctOption": 1,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     },
     {
         "question": "Which middleware is used in Express to parse JSON request bodies?",
         "options": ["body-parser", "json-parser", "express.json()", "request-parser"],
         "correctOption": 2,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     }
 ];
