@@ -3,60 +3,70 @@ export const pythonQuestions = [
         "question": "Which of the following is a mutable data type in Python?",
         "options": ["Tuple", "List", "String", "Integer"],
         "correctOption": 1,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     },
     {
         "question": "What does the `len()` function return in Python?",
         "options": ["Memory size of variable", "Number of elements in an iterable", "Data type of variable", "Index of an element"],
         "correctOption": 1,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     },
     {
         "question": "Which keyword is used to define a function in Python?",
         "options": ["func", "def", "define", "function"],
         "correctOption": 1,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     },
     {
         "question": "What is the output of `print(2**3)`?",
         "options": ["5", "6", "8", "9"],
         "correctOption": 2,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     },
     {
         "question": "Which data type is used to store a sequence of characters?",
         "options": ["int", "float", "str", "bool"],
         "correctOption": 2,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     },
     {
         "question": "What is the output of `type([])` in Python?",
         "options": ["dict", "list", "tuple", "set"],
         "correctOption": 1,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     },
     {
         "question": "Which function is used to read a file in Python?",
         "options": ["read()", "open()", "input()", "fetch()"],
         "correctOption": 1,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     },
     {
         "question": "Which statement is used to handle exceptions in Python?",
         "options": ["catch", "try-except", "throw", "raise"],
         "correctOption": 1,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     },
     {
         "question": "Which library is used for data analysis in Python?",
         "options": ["NumPy", "Pandas", "Matplotlib", "Seaborn"],
         "correctOption": 1,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     },
     {
         "question": "What will `bool([])` return?",
         "options": ["True", "False", "Error", "None"],
         "correctOption": 1,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     }
 ];
