@@ -3,60 +3,70 @@ export const sqlQuestions = [
         "question": "Which SQL command is used to retrieve data from a database?",
         "options": ["GET", "FETCH", "SELECT", "RETRIEVE"],
         "correctOption": 2,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     },
     {
         "question": "Which SQL clause is used to filter records?",
         "options": ["WHERE", "FILTER", "ORDER BY", "HAVING"],
         "correctOption": 0,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     },
     {
         "question": "Which SQL keyword is used to remove duplicate records?",
         "options": ["UNIQUE", "DISTINCT", "FILTER", "DELETE"],
         "correctOption": 1,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     },
     {
         "question": "Which command is used to update a record in SQL?",
         "options": ["CHANGE", "MODIFY", "UPDATE", "EDIT"],
         "correctOption": 2,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     },
     {
         "question": "Which SQL statement is used to insert data into a table?",
         "options": ["INSERT INTO", "ADD RECORD", "NEW ENTRY", "STORE"],
         "correctOption": 0,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     },
     {
         "question": "Which SQL function is used to find the total number of records in a table?",
         "options": ["SUM()", "COUNT()", "TOTAL()", "NUMBER()"],
         "correctOption": 1,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     },
     {
         "question": "What does the PRIMARY KEY constraint do in SQL?",
         "options": ["Allows duplicate values", "Ensures unique values", "Deletes duplicate records", "Creates a backup"],
         "correctOption": 1,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     },
     {
         "question": "Which SQL statement is used to delete all rows from a table without deleting the table structure?",
         "options": ["DROP", "DELETE", "REMOVE", "TRUNCATE"],
         "correctOption": 3,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     },
     {
         "question": "What is the purpose of the GROUP BY clause in SQL?",
         "options": ["To group identical values", "To filter data", "To sort data", "To update records"],
         "correctOption": 0,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     },
     {
         "question": "Which SQL command is used to combine the results of two queries?",
         "options": ["JOIN", "COMBINE", "MERGE", "UNION"],
         "correctOption": 3,
-        "points": 10
+        "points": 10,
+        "category": "technical"
     }
 ];
